+++ internal/api/vpn/customergateway.go
@@ -43,6 +43,9 @@ type CustomerGatewayRequest struct {
 	ForceEncap      bool   `json:"forceencap"`
 	SplitConnection bool   `json:"splitConnection"`
 	DPD             bool   `json:"dpd"`
+	CloudProvider   string `json:"cloud_provider,omitempty"`
+	Region          string `json:"region,omitempty"`
+	Project         string `json:"project,omitempty"`
 }
 
 // CustomerGatewayService provides VPN customer gateway API operations.